+++ src/Components/Apartment/AparmtentOrders.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from "react";
 import Axios from "axios";
 
-//  This page should display all orders that the apartment has
+//  This page should display all orders that the apartment has by grabbing the
+//  aparmentID from localStorage and doing a post request
 
 export const AparmtentOrders = () => {
   const [orders, setOrders] = useState([]);
