--- src/Login.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const Login = () => {
-  return <div>Login</div>;
-};
